backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
redis[hiredis]==5.0.1
httpx==0.27.0
anthropic==0.18.1
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-api-python-client==2.118.0
slack-sdk==3.27.1
aiohttp==3.9.3
websockets==12.0
